add (maybe) to inaccurate symbols and fix issue caused by it
Previously, we fixed unresovled symbols by not checking the size of
symbol. It causes issues for trying to resovle symbols in minidebug
because of the symbol cache.

Also it's hard to distinguish whether the symbol is accurate or not in
ui.

So we fix the issue and add (maybe) before the inaccurate symbols.