workaround: limit size of query for labels to avoid slow query
Querying labels is slow when database becomes large. Orignally, Labels
will query all of the records to get labels.

In the workaround, we limit the size of records. The side effect is
some old labels will not be shown if it's not stored recently. So we
bypass the value checks to let all the label be shown.