workaround: query schema instead of table for labels to avoid slow query
In the previous path we tried to use Limit() to speed up query. But
Limit() only effects to the output and not effects to the real database
query

So we use schema query instead. The side effect is it will get all the
labels, so the profileType filter has no effect then.